Update project refactoring it Directorys
