+++ pet_store_karate/pom.xml
@@ -5,15 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.example</groupId>
-<<<<<<< HEAD
     <artifactId>petstore</artifactId>
-=======
-<<<<<<< HEAD
-    <artifactId>petstore</artifactId>
-=======
-    <artifactId>request</artifactId>
->>>>>>> 72ce018b6e7123c8c8700d2ad772feb0850fd321
->>>>>>> 5f34302b906dd64e8b33513fd401816d94d1fbc2
     <version>1.0-SNAPSHOT</version>
 
     <properties>
@@ -74,12 +66,5 @@
             </plugin>
         </plugins>
     </build>
-<<<<<<< HEAD
 
-=======
-<<<<<<< HEAD
-
-=======
->>>>>>> 72ce018b6e7123c8c8700d2ad772feb0850fd321
->>>>>>> 5f34302b906dd64e8b33513fd401816d94d1fbc2
 </project>
