+++ pet_store_karate/src/test/java/karate/request/AddPets/addPets.java
@@ -1,4 +1,4 @@
-package karate.request;
+package karate.request.AddPets;
 
 import com.intuit.karate.junit5.Karate;
