+++ pet_store_karate/src/test/java/karate/request/DeletePet/deletePet.java
@@ -1,4 +1,4 @@
-package karate.request;
+package karate.request.DeletePet;
 
 import com.intuit.karate.junit5.Karate;
