+++ pet_store_karate/src/test/resources/karate/request/addPets/responsePost.json
