+++ pet_store_karate/src/test/resources/karate/request/deletePet/resposeErrorDelete.json
