+++ pet_store_karate/src/test/resources/karate/request/findPetId/responseErrorGet.json
