+++ pet_store_karate/src/test/resources/karate/request/updatePet/responseErrorPut.json
